<version> 0.3.9 <feature> fix: 被暂停的声音不应被自动移除

+++ Assets/GameTK/Feeling2DFramework/Modules_Sound/SoundModule.cs
@@ -149,8 +149,10 @@ namespace NJM {
                 var entity = array[i];
                 entity.FadeOut_Tick(dt);
                 if (!entity.IsPlaying()) {
-                    ctx.repo.Remove(entity.id);
-                    ctx.pool_entity.Return(entity);
+                    if (!entity.isPause) {
+                        ctx.repo.Remove(entity.id);
+                        ctx.pool_entity.Return(entity);
+                    }
                 } else {
                     if (entity.isFollowBelong) {
                         Vector2 pos = events.OnGetBelongPos(entity.belong);
